+++ addOns/pscanrulesAlpha/src/test/java/org/zaproxy/zap/extension/pscanrulesAlpha/UserControlledOpenRedirectScannerUnitTest.java
@@ -217,4 +217,35 @@ public class UserControlledOpenRedirectScannerUnitTest
         // Then
         assertThat(alertsRaised.size(), equalTo(0));
     }
+
+    @Test
+    public void
+            shouldNotRaiseAlertIfresponseIsRedirectAndParamIsOnlyMatchingProtocolOfLocationHeader() {
+        // Given
+        HttpMessage msg = createMessage();
+        TreeSet<HtmlParameter> params = new TreeSet<HtmlParameter>();
+        params.add(new HtmlParameter(HtmlParameter.Type.url, "place", "http"));
+        msg.setGetParams(params);
+        msg.getResponseHeader().setStatusCode(HttpStatusCode.MOVED_PERMANENTLY);
+        msg.getResponseHeader().setHeader(HttpHeader.LOCATION, "http://evil.com/xyz");
+        // When
+        rule.scanHttpResponseReceive(msg, -1, createSource(msg));
+        // Then
+        assertThat(alertsRaised.size(), equalTo(0));
+    }
+
+    @Test
+    public void shouldNotRaiseAlertIfResponseIsRedirectHasRelativeLocationHeader() {
+        // Given
+        HttpMessage msg = createMessage();
+        TreeSet<HtmlParameter> params = new TreeSet<HtmlParameter>();
+        params.add(new HtmlParameter(HtmlParameter.Type.url, "place", "/images"));
+        msg.setGetParams(params);
+        msg.getResponseHeader().setStatusCode(HttpStatusCode.MOVED_PERMANENTLY);
+        msg.getResponseHeader().setHeader(HttpHeader.LOCATION, "/images");
+        // When
+        rule.scanHttpResponseReceive(msg, -1, createSource(msg));
+        // Then
+        assertThat(alertsRaised.size(), equalTo(0));
+    }
 }